fastapi-backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
Authlib==1.4.1
bcrypt==4.2.1
boto3==1.36.19
botocore==1.36.19
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
frozenlist==1.5.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-speech==2.31.0
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.8.2
jiwer==3.1.0
jmespath==1.0.1
limits==4.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.6.0
multidict==6.1.0
numpy==2.2.3
openai==1.62.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
polling2==0.5.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyannote.core==5.0.0
pyannote.database==5.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.11.2
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
speechmatics-python==3.0.2
SQLAlchemy==2.0.38
starlette==0.45.3
tabulate==0.9.0
tenacity==8.2.3
toml==0.10.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
